Replace compiler with refactored version
