+++ rat/compiler.py
@@ -7,82 +7,332 @@ import jax
 import jax.numpy
 import jax.scipy.stats
 from typing import Iterable, List, Dict, Set, Tuple, Union
+from enum import Enum
+from dataclasses import dataclass, field
+from collections import defaultdict
+import pprint
+import itertools
 import warnings
 
+import pandas as pd
+
 from . import ast
-from . import variables
+from . import codegen_backends
 
 
 class CompileError(Exception):
-    def __init__(self, message, code_string: str, line_num: int = -1, column_num: int = -1):
+    def __init__(self, message, code_string: str = "", line_num: int = -1, column_num: int = -1):
         code_string = code_string.split("\n")[line_num] if code_string else ""
-        exception_message = f"An error occured while compiling the following line({line_num}:{column_num}):\n{code_string}\n{' ' * column_num + '^'}\n{message}"
+        if code_string:
+            exception_message = f"An error occurred while compiling the following line({line_num}:{column_num}):\n{code_string}\n{' ' * column_num + '^'}\n{message}"
+        else:
+            exception_message = f"An error occurred during compilation:\n{message}"
         super().__init__(exception_message)
+
+
+class CompileWarning(UserWarning):
+    def __init__(self, message, code_string: str = "", line_num: int = -1, column_num: int = -1):
+        code_string = code_string.split("\n")[line_num] if code_string else ""
+        if code_string:
+            warning_message = f"Compilation warning({line_num}:{column_num}):\n{code_string}\n{' ' * column_num + '^'}\n{message}"
+        else:
+            warning_message = f"Compilation warning:\n{message}"
+        super().__init__(warning_message)
+
+
+class VariableType(Enum):
+    DATA = 1
+    PARAM = 2
+    ASSIGNED_PARAM = 3
+
+
+@dataclass()
+class TableRecord:
+    """
+    A record within the SymbolTable
+
+    name: Tame of the variable
+    variable_type: The type of the variable(VariableType). Data, Param, or Assigned Param
+    subscripts: If subscripts exist, the "true" subscript set of the variable. All subscripts aliases effectively
+    point to one or more columns of the input dataframe. Those columns of the input dataframe are the true subscripts
+    of the variable. If the variable has 2 subscripts, it is stored as {(column_1, column_2), ...}
+    subscript_length: the length of the subscript, that is, how many subscripts are declared for the variable.
+    subscript_alias: This is the "fake" subscript names, declared by the user for the variable. These values are used
+    as column names of the base dataframe
+    constraint_lower: value of the lower constraint
+    constraint_upper: value of the upper constraint
+    base_df: the calculated base dataframe. This is generated by `SymbolTable.build_base_dataframes()` and shouldn't be set
+    by the programmer.
+    unconstrained_vector_start_index: For parameters, the start index of the unconstrained parameter vector. Generated
+    by `SymbolTable.build_base_dataframes()`.
+    unconstrained_vector_end_index: For parameters, the end index of the unconstrained parameter vector. Generated by
+    `SymbolTable.build_base_dataframes()`.
+    pad_needed: true if a padding is needed for the variable
+    """
+
+    name: str
+    variable_type: VariableType
+    subscripts: Set[Tuple[str]]  # tuple of (subscript_name, subscript_name, ...)
+    subscript_length: int = field(default=0)
+    subscript_alias: Tuple[str] = field(default=tuple())
+    constraint_lower: float = field(default=float("-inf"))
+    constraint_upper: float = field(default=float("inf"))
+    base_df: pd.DataFrame = field(default=None, init=False, repr=False)
+    unconstrained_vector_start_index: int = field(default=None, init=False)
+    unconstrained_vector_end_index: int = field(default=None, init=False)
+    pad_needed: bool = field(default=False, init=False)
+
+
+class SymbolTable:
+    def __init__(self, data_df):
+        """
+        symbol_dict: The internal dictionary that represents the symbol table. Key values are variable names.
+        unconstrained_param_count: the length of the required unconstrained parameter vector.
+        data_df: the input data dataframe
+        generated_subscript_dict: Dictionary which maps subscript keys to indices.
+        """
+        self.symbol_dict: Dict[str, TableRecord] = {}
+        self.unconstrained_param_count: int = 0  # length of the unconstrained parameter vector
+        self.data_df = data_df
+        self.generated_subscript_dict: Dict[str, numpy.ndarray] = {}
+        self.generated_subscript_count = 0
+        self.first_in_group_indicator: Dict[str, numpy.ndarray] = {}
+
+        self._unique_number = 0
+
+    def get_unique_number(self):
+        self._unique_number += 1
+        return self._unique_number
+
+    def upsert(
+        self,
+        variable_name: str,
+        variable_type: VariableType = None,
+        subscripts: Set[Tuple[str]] = None,
+        subscript_alias: Tuple[str] = tuple(),
+        constraint_lower: float = float("-inf"),
+        constraint_upper: float = float("inf"),
+    ):
+        """
+        Upsert(update or create) a record within the symbol table. The 6 arguments of this functions should be the only
+        fields of the record that the programmer should provide; other fields are generated automatically by
+        `build_base_dataframes()`.
+        """
+        if variable_name in self.symbol_dict:
+            # Update fields accordingly
+            record = self.symbol_dict[variable_name]
+            if variable_type:
+                record.variable_type = variable_type
+
+            if subscripts:
+                if record.subscript_length > 0:
+                    # check that the length of subscript operations all match
+                    for subscript_tuple in subscripts:
+                        if len(subscript_tuple) != record.subscript_length:
+                            raise ValueError(
+                                f"Internal error - symbol table update failed. Variable '{variable_name}' in table has {len(record.subscripts)} subscripts, but update request has {len(subscripts)}"
+                            )
+
+                record.subscript_length = len(tuple(subscripts)[0])
+                record.subscripts |= subscripts
+
+            if subscript_alias:
+                record.subscript_alias = subscript_alias
+
+            record.constraint_lower = max(record.constraint_lower, constraint_lower)
+            record.constraint_upper = min(record.constraint_upper, constraint_upper)
+
+        else:
+            # Insert if new
+            self.symbol_dict[variable_name] = TableRecord(
+                variable_name,
+                variable_type,
+                subscripts=subscripts if subscripts else set(),
+                subscript_length=len(tuple(subscripts)[0]) if subscripts else 0,
+                subscript_alias=subscript_alias,
+                constraint_lower=constraint_lower,
+                constraint_upper=constraint_upper,
+            )
+
+    def lookup(self, variable_name: str):
+        """
+        Dictionary indexing
+        """
+        return self.symbol_dict[variable_name]
+
+    def iter_records(self):
+        for name, record in self.symbol_dict.items():
+            yield name, record
+
+    def build_base_dataframes(self, data_df: pd.DataFrame):
+        """
+        Builds the base dataframes for parameters from the current symbol table.
+        Also resets any generated subscript indices
+        """
+        self.generated_subscript_dict = {}
+        self.generated_subscript_count = 0
+
+        current_index = 0
+        for variable_name, record in self.symbol_dict.items():
+            if not record.variable_type:
+                error_msg = f"Fatal internal error - variable type information for '{variable_name}' not present within the symbol table. Aborting compilation."
+                raise CompileError(error_msg)
+
+            # Data variables don't need a base df; the input df is its base dataframe.
+            if record.variable_type == VariableType.DATA:
+                continue
+
+            if record.subscript_length > 0:
+                base_df = pd.DataFrame()
+
+                # Since the subscripts in the symbol table are all resolved to denote columns in the input dataframe,
+                # we can directly pull out columns from the input dataframe.
+                for subscript_tuple in record.subscripts:
+                    try:
+                        df = data_df.loc[:, subscript_tuple]
+                    except KeyError as e:
+                        raise CompileError(
+                            f"Internal Error - dataframe build failed. Could not index one or more columns in {subscript_tuple} from the data dataframe."
+                        ) from e
+
+                    # rename base dataframe's columns to use the subscript alias
+                    df.columns = tuple(record.subscript_alias)
+                    base_df = pd.concat([base_df, df]).drop_duplicates().sort_values(list(df.columns)).reset_index(drop=True)
+
+                # For parameters, allocate space on the unconstrained parameter vector and record indices
+                if record.variable_type == VariableType.PARAM:
+                    nrows = base_df.shape[0]
+                    #base_df["__index"] = pd.Series(range(current_index, current_index + nrows))
+                    record.unconstrained_vector_start_index = current_index
+                    record.unconstrained_vector_end_index = current_index + nrows - 1
+                    current_index += nrows
+
+                record.base_df = base_df
+
+            else:
+                if record.variable_type == VariableType.PARAM:
+                    record.unconstrained_vector_start_index = current_index
+                    record.unconstrained_vector_end_index = current_index
+                    current_index += 1
+
+        self.unconstrained_param_count = current_index
+
+    def get_subscript_key(
+        self,
+        primary_variable_name: str,
+        primary_subscript_names: Tuple[str],
+        target_variable_name: str,
+        target_subscript_names: Tuple[str],
+        target_shift_amounts: Tuple[int],
+    ):
+        """
+        Find the subscript indices for (target variable, subscript, shift) given its primary variable and its declared subscript
+        For example:
+        param_a[s1, s2] = param_b[s1]
+
+        Procedure:
+        1. filter out base_df of param_a so that only rows containing s1 and s2 exists.
+        2. filter out base_df of param_b so that only rows containing s1 exists.
+        3. apply shifts to dataframe (1), creating new columns on dataframe 1
+        3. merge dataframe (1) with dataframe (2) along s1, keeping indices from dataframe (2)
+
+        """
+        primary_record = self.lookup(primary_variable_name)
+        target_record = self.lookup(target_variable_name)
+
+        if primary_record.variable_type == VariableType.DATA:
+            primary_base_df = self.data_df.copy()
+
+        else:
+            primary_base_df = primary_record.base_df.copy()
+
+            for index, subscript in enumerate(primary_subscript_names):
+                base_df_subscript_name = primary_record.subscript_alias[index]
+                if subscript == base_df_subscript_name:
+                    continue
+                else:
+                    subset_df = self.data_df.loc[:, [subscript]].drop_duplicates()
+                    subset_df.columns = (base_df_subscript_name,)
+                    primary_base_df = pd.merge(primary_base_df, subset_df, on=base_df_subscript_name, how="inner")
+
+        primary_base_df = primary_base_df[list(target_subscript_names)]
+
+        shift_subscripts = []
+        shift_values = []
+        grouping_subscripts = []
+        for column, shift in zip(target_subscript_names, target_shift_amounts):
+            if shift == 0:
+                grouping_subscripts.append(column)
+            else:
+                shift_subscripts.append(column)
+                shift_values.append(shift)
+
+        if len(grouping_subscripts) > 0:
+            grouped_df = primary_base_df.groupby(grouping_subscripts)
+        else:
+            grouped_df = primary_base_df
+
+        for column, shift in zip(shift_subscripts, shift_values):
+            shifted_column = grouped_df[column].shift(shift).reset_index(drop=True)
+            primary_base_df[column] = shifted_column
+
+        #target_base_df = target_record.base_df.copy().loc[:, target_record.base_df.columns != "__index"]
+        target_base_df = target_record.base_df.copy()
+        target_base_df.columns = list(target_subscript_names)
+
+        target_base_df["__in_dataframe_index"] = pd.Series(range(target_base_df.shape[0]))
+
+        key_name = f"subscript__{self.generated_subscript_count}"
+        self.generated_subscript_count += 1
+        output_df = pd.merge(
+            primary_base_df[list(target_subscript_names)], target_base_df, on=target_subscript_names, how="left"
+        )
+        if sum([1 if x != 0 else 0 for x in shift_values]) > 1:
+            self.first_in_group_indicator[target_variable_name] = (
+                ~output_df.duplicated(subset=grouping_subscripts)).to_numpy()
+
+        output_df["__in_dataframe_index"] = output_df["__in_dataframe_index"].fillna(target_base_df.shape[0])
+        self.generated_subscript_dict[key_name] = output_df["__in_dataframe_index"].to_numpy(dtype=int)
+
+        return key_name
+
+    def __str__(self):
+        return pprint.pformat(self.symbol_dict)
 
 
 class Compiler:
     data_df: pandas.DataFrame
-    expr_tree_list: List[ops.Expr]
+    expr_tree_list: List[ast.Expr]
     model_code_string: str
-    # There are a couple places internally here where it is handy to have a unique
-    # identifier. We can use increasing integers to do this
-    _unique_identifier: int
 
-    def __init__(self, data_df: pandas.DataFrame, expr_tree_list: List[ops.Expr], model_code_string: str = ""):
-        """
-        The rat compiler receives the data dataframe and parsed expression trees and generates Python code which is used
-        to declare the target logpdf for inference. During compilation, a total of 2 passes are done through the
-        expression tree, excluding 2 separate topological sort operations for identifying compilation order.
-        The first pass is based on reverse dependency order and resolves subscripts, while the second pass runs in
-        standard dependence order and does the actual codegen/linking with ops.
-
-        In a high level, Compiler does the following operations:
-
-        1. generate reverse dependency graph
-        2-1. generate base_df(reference df) for every subscripted parameter
-        2-2. generate variable.Subscript() for each variable/data
-        3. generate canonical dependency graph
-        4-1. generate variable.Param/AssignedParam/Data and variable.SubscriptUse()
-        4-2. generate python source for model
-        """
+    def __init__(self, data_df: pandas.DataFrame, expr_tree_list: List[ast.Expr], model_code_string: str = ""):
         self.data_df = data_df
+        self.data_df_columns = list(data_df.columns)
         self.expr_tree_list = expr_tree_list
         self.model_code_string = model_code_string
-        self._unique_identifier = 0
+        self.symbol_table: SymbolTable = None
+        self.base_df: Dict[str, pd.DataFrame] = {}
+        self.generated_code = ""
 
-        # these dataframes that go into variables.Subscript. Each row is a unique combination of indexes
-        self.parameter_base_df: Dict[str, pandas.DataFrame] = {}
-
-        # these dataframes that go into variables.Subscript
-        self.data_base_df: Dict[str, pandas.DataFrame] = {}
-
-        # this is the dictionary that keeps track of what names the parameter subscripts were defined as
-        self.parameter_subscript_names: Dict[str, List[Set[str]]] = {}
-
-        # this set keep track of variable names that are not parameters, i.e. assigned by assignment
-        self.assigned_parameter_keys: Set[str] = set()
-
-        self.data_variables: Dict[str, variables.Data] = {}
-        self.parameter_variables: Dict[str, variables.Param] = {}
-        self.assigned_parameter_variables: Dict[str, variables.AssignedParam] = {}
-        self.subscript_use_variables: Dict[str, variables.SubscriptUse] = {}
-        self.variable_subscripts: Dict[str, variables.Subscript] = {}
-        self.first_in_group_indicators: Dict[str, numpy.ndarray] = {}
-
-    def _get_unique_identifier(self) -> str:
+    def _get_primary_symbol_from_statement(self, top_expr):
         """
-        Return an as-yet un-returned string. This is useful for generating unique identifiers
-        for variables
+        Get the primary symbol in a statement. This assumes that the statement has
+        only one primary symbol
         """
-        self._unique_identifier += 1
-        return repr(self._unique_identifier)
+        for primary_symbol in ast.search_tree(top_expr, ast.PrimeableExpr):
+            if primary_symbol.prime:
+                break
+        else:
+            msg = f"Internal compiler error. No primary variable found"
+            raise CompileError(msg, self.model_code_string, top_expr.line_index, top_expr.column_index)
+        return primary_symbol
 
     def _identify_primary_symbols(self):
         # figure out which symbols will have dataframes
         has_dataframe = set()
         for top_expr in self.expr_tree_list:
-            for primeable_symbol in ops.search_tree(top_expr, ops.PrimeableExpr):
-                if isinstance(primeable_symbol, ops.Data) or primeable_symbol.subscript is not None:
+            for primeable_symbol in ast.search_tree(top_expr, ast.PrimeableExpr):
+                if isinstance(primeable_symbol, ast.Data) or primeable_symbol.subscript is not None:
                     has_dataframe.add(primeable_symbol.get_key())
 
         for top_expr in self.expr_tree_list:
@@ -92,9 +342,9 @@ class Compiler:
             # 3. If there is no marked primary variable, then all variables with dataframes are treated as prime.
             # 4. If there are no variables with dataframes, the leftmost one is the primary one
             # 5. It is an error if no primary variable can be identified
-            primary_symbol: ops.PrimeableExpr = None
+            primary_symbol: ast.PrimeableExpr = None
             # Rule 2
-            for primeable_symbol in ops.search_tree(top_expr, ops.PrimeableExpr):
+            for primeable_symbol in ast.search_tree(top_expr, ast.PrimeableExpr):
                 if primeable_symbol.prime:
                     if primary_symbol is None:
                         primary_symbol = primeable_symbol
@@ -104,7 +354,7 @@ class Compiler:
 
             # Rule 3
             if primary_symbol is None:
-                for primeable_symbol in ops.search_tree(top_expr, ops.PrimeableExpr):
+                for primeable_symbol in ast.search_tree(top_expr, ast.PrimeableExpr):
                     primeable_key = primeable_symbol.get_key()
 
                     if primeable_key in has_dataframe:
@@ -121,7 +371,7 @@ class Compiler:
 
             # Rule 4
             if primary_symbol is None:
-                for primeable_symbol in ops.search_tree(top_expr, ops.PrimeableExpr):
+                for primeable_symbol in ast.search_tree(top_expr, ast.PrimeableExpr):
                     primary_symbol = primeable_symbol
                     break
 
@@ -133,575 +383,262 @@ class Compiler:
             # Mark the primary symbol if it wasn't already
             primary_symbol.prime = True
 
-    def _get_primary_symbol_from_statement(self, top_expr):
+    def build_symbol_table(self):
         """
-        Get the primary symbol in a statement. This assumes that the statement has
-        only one primary symbol
+        Builds the "symbol table", which holds information for all variables in the model.
         """
-        for primary_symbol in ops.search_tree(top_expr, ops.PrimeableExpr):
-            if primary_symbol.prime:
-                break
-        else:
-            msg = f"Internal compiler error. No primary variable found"
-            raise CompileError(msg, self.model_code_string, top_expr.line_index, top_expr.column_index)
-        return primary_symbol
-
-    def _build_base_df(self):
-        """
-        First code pass. This function must receive the *reversed* dependence graph, or RHS | LHS where edges go from
-        RHS to LHS. This function will loop over the expression trees once, and write the parameter base dataframes into
-        the class attribute self. parameter_base_df.
-
-        The core two classes related to subscripts are `variables.Subscript` and `variables.SubscriptUse`. In short,
-        `Subscript` denotes the dataframe where the variable is used in LHS. It's used as the basis for resolving
-        subscripts on its RHS.
-        Meanwhile, `SubscriptUse` is what's given to Param in RHS. It's inferred from the LHS variable's `Subscript`,
-        which is indexed according to the RHS subscript.
-
-        1. generate base_df for every LHS variable
-        2. identify every subscript scheme for all subscripted variables
-        """
-
-        # Make sure that all secondary parameter uses precede primary uses (or throw an error)
-        N_lines = len(self.expr_tree_list)
-        for i, top_expr in enumerate(self.expr_tree_list):
-            primary_symbol = self._get_primary_symbol_from_statement(top_expr)
-
-            primary_key = primary_symbol.get_key()
-
-            # There aren't any restrictions on primary/secondary data
-            if isinstance(primary_symbol, ops.Data):
-                continue
-
-            for j in range(i + 1, N_lines):
-                for secondary_symbol in ops.search_tree(self.expr_tree_list[j], ops.Param):
-                    secondary_key = secondary_symbol.get_key()
-
-                    if primary_key == secondary_key and not secondary_symbol.prime:
-                        msg = f"Primary variable {primary_symbol.get_key()} used on line {i} but then referenced as non-prime on line {j}. The primed uses must come last"
-                        raise CompileError(msg, self.model_code_string, primary_symbol.line_index, primary_symbol.column_index)
+        self.symbol_table = SymbolTable(self.data_df)
 
         for top_expr in self.expr_tree_list:
-            primary_symbol = self._get_primary_symbol_from_statement(top_expr)
-            # Identify the primary dataframe if there is one and resolve naming issues
-            primary_key = primary_symbol.get_key()
+            primary_variable = self._get_primary_symbol_from_statement(top_expr)
+            primary_variable_key = primary_variable.get_key()
 
-            match primary_symbol:
-                case ops.Data():
-                    # If the primary primary_symbol is data, the reference dataframe is the data dataframe
-                    primary_df = self.data_df
-                # TODO: This isn't implemented but might make sense
-                # case ops.Subscript():
-                # If the primary primary_symbol is a subscript, the reference dataframe is the data dataframe
-                #    primary_df = self.data_df
-                case ops.Param():
-                    # For parameters
-                    # 1. A subscripted primary parameter primary_symbol must have a dataframe
-                    # 2. If there is no subscript, then:
-                    #    A. If a dataframe exists, then column names must be uniquely defined
-                    #    B. If no dataframe exists, the parameter is a scalar
-                    # Since the lines are evaluated in a topological order, the rows of the primary dataframe are known at this point.
-                    if primary_key in self.parameter_base_df:
-                        primary_df = self.parameter_base_df[primary_key].copy()
+            if isinstance(primary_variable, ast.Data):
+                self.symbol_table.upsert(variable_name=primary_variable.name, variable_type=VariableType.DATA)
+            elif isinstance(primary_variable, ast.Param):
+                self.symbol_table.upsert(variable_name=primary_variable.name, variable_type=VariableType.PARAM)
 
-                        # If the primary primary_symbol is subscripted, should be a dataframe
-                        if primary_symbol.subscript:
-                            # Handle subscript renaming
-                            try:
-                                primary_df.columns = tuple(primary_symbol.subscript.get_key())
-                            except ValueError as e:
-                                # if usage subscript length and declared subscript length do not match, raise exception
-                                msg = f"Rename {primary_symbol.code()} not possible, {len(primary_df.columns)} required names but {len(primary_symbol.subscript.get_key())} given."
-                                raise CompileError(
-                                    msg, self.model_code_string, primary_symbol.line_index, primary_symbol.column_index
-                                ) from e
+            subscript_aliases: Dict[str, Set] = {}
+            # this dict stores any subscript aliases of the primary dataframe.
 
-                            self.parameter_base_df[primary_key].columns = primary_df.columns
-                        else:
-                            for n, in_use_subscripts in enumerate(self.parameter_subscript_names[primary_key]):
-                                if len(in_use_subscripts) > 1:
-                                    msg = f"Subscript(s) of {primary_key} must be renamed, the name for subscript {n} is ambiguous ({in_use_subscripts})"
-                                    raise CompileError(msg, self.model_code_string, primary_symbol.line_index, primary_symbol.column_index)
+            # find the primary variable and its subscript aliases
+            match primary_variable:
+                case ast.Data():
+                    # these are the allowed subscript names in scope
+                    allowed_subscript_names = tuple(self.data_df.columns)
+                case ast.Param():
+                    try:
+                        primary_variable_record = self.symbol_table.lookup(primary_variable_key)
+                    except KeyError:
+                        msg = f"Primary variable {primary_variable_key} does not exist on the symbol table. Did you define the variable before using it as a primary variable?"
+                        raise CompileError(msg, self.model_code_string, primary_variable.line_index, primary_variable.column_index)
                     else:
-                        primary_df = None
-                case _:
-                    msg = f"Expected a parameter or data primary_symbol but got type {primary_symbol.__class__.__name__}"
-                    raise CompileError(msg, self.model_code_string, primary_symbol.line_index, primary_symbol.column_index)
+                        if primary_variable_record.subscript_length > 0:
+                            # these are the allowed subscript names in scope
 
-            # Find every other parameter in the line and attempt to construct data
-            # frames for the ones with subscripts
-            for param in ops.search_tree(top_expr, ops.Param):
+                            if primary_variable.subscript:
+                                for index, subscript_column in enumerate(primary_variable.subscript.names):
+                                    if subscript_column.name not in self.data_df_columns:
+                                        subscript_aliases[subscript_column.name] = {
+                                            subscript_tuple[index] for subscript_tuple in primary_variable_record.subscripts
+                                        }
+
+                                self.symbol_table.upsert(
+                                    primary_variable_key, subscript_alias=tuple([x.name for x in primary_variable.subscript.names])
+                                )
+
+                            else:
+                                for index, subscript_column in enumerate(primary_variable_record.subscript_alias):
+                                    if subscript_column not in self.data_df_columns:
+                                        # If there's a subscript name that's not present in the input dataframe, this means
+                                        # it's a subscript "alias", which points to the "true" subscripts that's already on
+                                        # the symbol table. We look up its "true" subscripts and keep them in the scope of
+                                        # this line
+                                        subscript_aliases[subscript_column] = {
+                                            subscript_tuple[index] for subscript_tuple in primary_variable_record.subscripts
+                                        }
+                            allowed_subscript_names = primary_variable_record.subscript_alias
+                        else:
+                            # these are the allowed subscript names in scope
+                            allowed_subscript_names = tuple()
+
+                case _:
+                    msg = f"Expected a parameter or data primary variable but got type {primary_variable.__class__.__name__}"
+                    raise CompileError(msg, self.model_code_string, primary_variable.line_index, primary_variable.column_index)
+
+            # If it's an assignment, we save the type of the LHS variable to be an assigned parameter
+            if isinstance(top_expr, ast.Assignment):
+                self.symbol_table.upsert(variable_name=top_expr.lhs.name, variable_type=VariableType.ASSIGNED_PARAM)
+
+            # iterate through the remaining non-primary variables, resolving subscripts according to primary's subscripts
+            for param in ast.search_tree(top_expr, ast.Param):
                 param_key = param.get_key()
 
-                # no need to check the primary symbol
-                if param_key == primary_key:
-                    continue
+                # If there are constraints, we evaluate their values and store them on the symbol table
+                try:
+                    lower_constraint_evaluator = codegen_backends.BaseVisitor(self.symbol_table)
+                    param.lower.accept(lower_constraint_evaluator)
+                    lower_constraint_value = float(eval(lower_constraint_evaluator.expression_string))
 
-                # if the symbol is not subscripted, nothing more needs to be done
+                    upper_constraint_evaluator = codegen_backends.BaseVisitor(self.symbol_table)
+                    param.upper.accept(upper_constraint_evaluator)
+                    upper_constraint_value = float(eval(upper_constraint_evaluator.expression_string))
+                except Exception as e:
+                    error_msg = f"Failed evaluating constraints for parameter {param_key}"
+                    raise CompileError(error_msg, self.model_code_string, primary_variable.line_index, primary_variable.column_index) from e
+                else:
+                    self.symbol_table.upsert(
+                        variable_name=param_key, constraint_lower=lower_constraint_value, constraint_upper=upper_constraint_value
+                    )
+                ###########
+
+                # The following operations are done for subscripts only
                 if not param.subscript:
                     continue
 
-                # get the subscripted columns from primary df
-                try:
-                    use_df = primary_df.loc[:, param.subscript.get_key()]
-                except KeyError as e:
-                    # this means a subscript referenced does not exist in the primary dataframe
-                    msg = f"Couldn't index RHS variable {param_key}'s declared subscript{param.subscript.get_key()} from LHS base df, which has subscripts {tuple(primary_df.columns)}."
-                    raise CompileError(msg, self.model_code_string, param.line_index, param.column_index) from e
-                except AttributeError as e:
-                    # this means LHS has no base dataframe but a subscript was attempted from RHS
-                    msg = f"Variable {param_key} attempted to subscript using {param.subscript.get_key()}, but primary variable {primary_key} has no subscripts!"
-                    raise CompileError(msg, self.model_code_string, param.line_index, param.column_index) from e
-
-                if param_key in self.parameter_base_df:
-                    # if we already have a base_df registered, try merging them
-                    try:
-                        use_df.columns = tuple(self.parameter_base_df[param_key])
-                    except ValueError as e:
-                        msg = f"Subscript length mismatch: variable on RHS side {param_key} has declared subscript of length {len(param_key)}, but has been used with a subscript of length {len(self.parameter_base_df[param_key])}"
-                        raise CompileError(msg, self.model_code_string, param.line_index, param.column_index) from e
-
-                    self.parameter_base_df[param_key] = (
-                        pandas.concat([self.parameter_base_df[param_key], use_df])
-                        .drop_duplicates()
-                        .sort_values(list(use_df.columns))
-                        .reset_index(drop=True)
-                    )
-
-                    # keep track of subscript names for each position
-                    for n, subscript in enumerate(param.subscript.get_key()):
-                        self.parameter_subscript_names[param_key][n].add(subscript)
-
-                else:  # or generate base df for RHS variable if it doesn't exist
-                    # this list of sets keep track of the aliases of each subscript position
-                    self.parameter_subscript_names[param_key] = [set() for _ in range((len(param.subscript.get_key())))]
-                    for n, subscript in enumerate(param.subscript.get_key()):
-                        self.parameter_subscript_names[param_key][n].add(subscript)
-                    self.parameter_base_df[param_key] = use_df.drop_duplicates().sort_values(list(use_df.columns)).reset_index(drop=True)
-
-            # TODO: This check will need to get fancier when there are multiple dataframes
-            # Find every Data in the line and make sure that they all have the necessary subscripts
-            for data in ops.search_tree(top_expr, ops.Data):
-                data_key = data.get_key()
-
-                self.data_base_df[data_key] = self.data_df
-
-    def _build_variables(self):
-        """
-        The goal in the second pass is
-        1. Generate variable.Param, variable.AssignedParam, variable.Data for each respective variable/data
-        2. Generate variable.SubscriptUse and map them to variable.Param/AssignedParam. They are used to resolve RHS
-        subscripts from the LHS base_df
-
-        The expression trees must be sorted in canonical topological order(LHS | RHS) in order to run succesfully.
-        """
-
-        # Parameters cannot be used after they are assigned
-        N_lines = len(self.expr_tree_list)
-        for i, top_expr in enumerate(self.expr_tree_list):
-            match top_expr:
-                case ops.Assignment(lhs):
-                    for j in range(i + 1, N_lines):
-                        lhs_key = lhs.get_key()
-                        for symbol in ops.search_tree(self.expr_tree_list[j], ops.Param):
-                            symbol_key = symbol.get_key()
-                            if symbol_key == lhs_key:
-                                msg = f"Parameter {lhs.get_key()} is assigned on line {i} but used on line {j}. A variable cannot be used after it is assigned"
-                                raise CompileError(msg, self.model_code_string, lhs.line_index, lhs.column_index)
-
-        ordered_expr_trees = list(reversed(self.expr_tree_list))
-
-        # Create variable objects
-        for top_expr in ordered_expr_trees:
-            # if top_expr is an assignment or a match statement, pull out the left hand side and treat that specially
-            match top_expr:
-                case ops.Assignment(lhs, remainder):
-                    lhs_key = lhs.get_key()
-
-                    if not isinstance(lhs, ops.Param):
-                        msg = f"Left hand side for assignment statement must be a parameter instead of {lhs.__class__.__name__}"
-                        raise CompileError(msg, self.model_code_string, lhs.line_index, lhs.column_index)
-
-                    # create the variable if necessary
-                    if lhs_key not in self.assigned_parameter_variables:
-                        # if the underlying variable has a dataframe, there will be an entry in variable_subscripts
-                        if lhs_key in self.variable_subscripts:
-                            lhs_variable = variables.AssignedParam(lhs, remainder, self.variable_subscripts[lhs_key])
-                        else:
-                            lhs_variable = variables.AssignedParam(lhs, remainder)
-
-                    # if this variable is stored in parameter_variables, remove the entry there
-                    if lhs_key in self.parameter_variables:
-                        self.parameter_variables.pop(lhs_key)
-
-                    self.assigned_parameter_variables[lhs_key] = lhs_variable
-                case remainder:
-                    pass
-
-            # remainder at this point is the part of top_expr that is not the left hand side of an assignment
-            for symbol in ops.search_tree(remainder, ops.Data, ops.Param):
-                symbol_key = symbol.get_key()
-
-                match symbol:
-                    case ops.Data():
-                        # create the variable if necessary
-                        if symbol_key not in self.data_variables:
-                            self.data_variables[symbol_key] = variables.Data(
-                                name=symbol_key,
-                                series=self.data_base_df[symbol_key].loc[:, symbol_key],
-                                subscript=self.variable_subscripts[symbol_key],
-                            )
-                    case ops.Param():
-                        # create the variable if necessary
-                        if symbol_key not in self.assigned_parameter_variables and symbol_key not in self.parameter_variables:
-                            # if the underlying variable has a dataframe, there will be an entry in variable_subscripts
-                            if symbol_key in self.variable_subscripts:
-                                variable = variables.Param(name=symbol_key, subscript=self.variable_subscripts[symbol_key])
-                            else:
-                                variable = variables.Param(name=symbol_key)
-
-                            self.parameter_variables[symbol_key] = variable
-                    case _:
-                        msg = f"Internal compiler error"
-                        raise CompileError(msg, self.model_code_string, lhs.line_index, lhs.column_index)
-
-        # Handle constraints
-        for top_expr in ordered_expr_trees:
-            for parameter_symbol in ops.search_tree(top_expr, ops.Param):
-                parameter_key = parameter_symbol.get_key()
-
-                try:  # evaluate constraint expressions
-                    lower_expr = parameter_symbol.lower
-                    upper_expr = parameter_symbol.upper
-
-                    lower_constraint_value = float(eval(lower_expr.code())) if lower_expr is not None else None
-                    upper_constraint_value = float(eval(upper_expr.code())) if upper_expr is not None else None
-                except Exception as e:
-                    msg = f"Error when evaluating constraints for {parameter_key}. Constraint expressions must be able to be evaluated at compile time."
-                    raise CompileError(msg, self.model_code_string, parameter_symbol.line_index, parameter_symbol.column_index) from e
-
-                # we only pay attention to constraints on actual parameters
-                if parameter_key in self.parameter_variables:
-                    parameter_variable = self.parameter_variables[parameter_key]
-
-                    # apply constraints
-                    try:
-                        parameter_variable.set_constraints(lower_constraint_value, upper_constraint_value)
-                    except Exception as e:
-                        msg = f"Constraints of {parameter_key} have already been set"
-                        raise CompileError(msg, self.model_code_string, parameter_symbol.line_index, parameter_symbol.column_index)
-
-        # Link ops back to variables
-        for top_expr in ordered_expr_trees:
-            for symbol in ops.search_tree(top_expr, ops.Data, ops.Param):
-                symbol_key = symbol.get_key()
-
-                match symbol:
-                    case ops.Data():
-                        # point symbol to appropriate variable
-                        if symbol_key not in self.data_variables:
-                            msg = f"Internal compiler error: Could not find data variable {symbol_key}"
-                            raise CompileError(msg, self.model_code_string, symbol.line_index, symbol.column_index)
-
-                        symbol.variable = self.data_variables[symbol_key]
-
-                    case ops.Param():
-                        # If this symbol has a dataframe, then it must be left joinable to the primary dataframe
-                        if symbol_key in self.variable_subscripts:
-                            self.variable_subscripts[symbol_key]
-
-                        # point symbol to appropriate variable
-                        if symbol_key in self.assigned_parameter_variables:
-                            symbol.variable = self.assigned_parameter_variables[symbol_key]
-                        elif symbol_key in self.parameter_variables:
-                            symbol.variable = self.parameter_variables[symbol_key]
-                        else:
-                            msg = f"Internal compiler error: Could not find parameter variable {symbol_key}"
-                            raise CompileError(msg, self.model_code_string, symbol.line_index, symbol.column_index)
-
-        # Build SubscriptUse variables for every time a variable is subscripted
-        for top_expr in ordered_expr_trees:
-            # At this point there should only be one primary variable per line
-            primary_op = self._get_primary_symbol_from_statement(top_expr)
-
-            primary_key = primary_op.get_key()
-
-            for symbol in ops.search_tree(top_expr, ops.Data, ops.Param):
-                # For each data op in the line (excluding left-hand-sides):
-                # 1. Make sure that any dataframe attached to this symbol can be joined to the primary dataframe
-                # 2. Link the symbol to the appropriate variable
-                # 3. Build necessary SubscriptUse objects
-                symbol_key = symbol.get_key()
-
-                # resolve subscripts
-                if symbol.subscript is not None:
-                    primary_df = self.variable_subscripts[primary_key].base_df
-
-                    # evaluate shift amounts
-                    symbol.subscript.shifts = [int(eval(expr.code())) if expr else None for expr in symbol.subscript.shifts]
-
-                    # incorporate shifts into variable.Subscript
-                    self.variable_subscripts[symbol_key].incorporate_shifts(symbol.subscript.shifts)
-
-                    # extract the subscripts from the base df and link them into variable.SubscriptUse
-                    # TODO: I'm suspicious we need some sort of error checking on this giant statement -- like
-                    # I'm not sure why we expect primary_df to have the symbol subscripts here
-                    variable_subscript_use = variables.SubscriptUse(
-                        names=symbol.subscript.get_key(),
-                        df=primary_df.loc[:, symbol.subscript.get_key()],
-                        subscript=self.variable_subscripts[symbol_key],
-                        unique_id=self._get_unique_identifier(),
-                        shifts=symbol.subscript.shifts,
-                    )
-
-                    try:
-                        variable_subscript_use.to_numpy()
-                    except Exception as e:
-                        msg = f"Could not uniquely join rows of dataframe of {symbol_key} into {primary_key} on {symbol_key}"
-                        raise CompileError(msg, self.model_code_string, symbol.line_index, symbol.column_index)
-
-                    # link the created variable.SubscriptUse into ops.Param
-                    symbol.subscript.variable = variable_subscript_use
-                    self.subscript_use_variables[variable_subscript_use.code()] = variable_subscript_use
-
-                    # quickly check to make sure subscripts don't exist for Data
-                    # for subexpr in ops.search_tree(top_expr, ops.Data):
-                    #    subexpr_key = subexpr.get_key()
-                    #    if subexpr.subscript:
-                    #        raise CompileError(
-                    #            f"Indexing on data variables ({subexpr_key}) not supported",
-                    #            self.model_code_string,
-                    #            subexpr.line_index,
-                    #            subexpr.column_index,
-                    #        )
-
-            # self.line_functions.append(line_function)
-            logging.debug(f"data: {list(self.data_variables.keys())}")
-            logging.debug(f"parameters: {list(self.parameter_variables.keys())}")
-            logging.debug(f"assigned parameters {list(self.assigned_parameter_variables.keys())}")
-
-    def _prepare_for_scans(self):
-        # If a variable is assigned on a given line, mark uses so that the code
-        # generation for the scan can handle this case correctly
-        for top_expr in self.expr_tree_list:
-            match top_expr:
-                case ops.Assignment(lhs, rhs):
-                    # Assume already that the left hand side is a Param
-                    assigned_key = lhs.get_key()
-
-                    # The left hand side is written by the assignment but will be updated
-                    # after the scan is complete and should not be marked
-                    for symbol in ops.search_tree(rhs, ops.Param):
-                        symbol_key = symbol.get_key()
-
-                        if assigned_key == symbol_key:
-                            symbol.assigned_by_scan = True
-
-                            if symbol.subscript is None or all(shift == 0 for shift in symbol.subscript.shifts):
-                                msg = f"Recursively assigning {symbol_key} requires a shifted subscript on the right hand side reference"
-                                raise CompileError(msg, self.model_code_string, symbol.line_index, symbol.column_index)
-
-                            shifts = symbol.subscript.shifts
-
-                            if sum(shift != 0 for shift in shifts) != 1:
-                                msg = "Exactly one (no more, no less) subscripts can be shifted in a recursively assigned variable"
-                                raise CompileError(msg, self.model_code_string, symbol.line_index, symbol.column_index)
-
-                            if shifts[-1] == 0:
-                                msg = "Only the right-most subscript of a recursively assigned variable can shifted"
-                                raise CompileError(msg, self.model_code_string, symbol.line_index, symbol.column_index)
-
-                            if any(shift < 0 for shift in shifts):
-                                msg = "All shifts in a recursively assigned variable must be greater than or equal to zero"
-                                raise CompileError(msg, self.model_code_string, symbol.line_index, symbol.column_index)
-
-                            # Generate the first in group indicators used to mask the scan carries
-                            self.first_in_group_indicators[assigned_key] = symbol.subscript.variable.get_first_in_group_indicators()
-
-    def _generate_python_source(self):
-        """
-        Return python source containing functions for constraining and transforming variables and
-        evaluating all the log densities
-        """
-        # Get parameter dimensions so can build individual arguments from
-        # unconstrained vector
-        unconstrained_parameter_size = 0
-        parameter_names = []
-        parameter_offsets = []
-        parameter_sizes = []
-        for name, parameter in self.parameter_variables.items():
-            parameter_names.append(name)
-            parameter_offsets.append(unconstrained_parameter_size)
-            parameter_size = parameter.size()
-            parameter_sizes.append(parameter_size)
-
-            if parameter_size is not None:
-                unconstrained_parameter_size += parameter_size
-            else:
-                unconstrained_parameter_size += 1
-
-        # Generate code for unconstraining and transforming parameters
-        code = (
-            f"import rat.constraints\n"
-            f"import rat.math\n"
-            f"import jax\n"
-            f"\nunconstrained_parameter_size = {unconstrained_parameter_size}\n"
-            f"\ndef constrain_parameters(unconstrained_parameter_vector, pad=True):\n"
-            f"    unconstrained_parameters = {{}}\n"
-            f"    parameters = {{}}\n"
-            f"    total = 0.0\n"
-        )
-
-        # Constrain parameters
-        for name, offset, size in zip(parameter_names, parameter_offsets, parameter_sizes):
-            unconstrained_reference = f"unconstrained_parameters['{name}']"
-            constrained_reference = f"parameters['{name}']"
-            code += f"\n    # {name}\n"
-            if size is not None:
-                code += f"    {unconstrained_reference} = unconstrained_parameter_vector[..., {offset} : {offset + size}]\n"
-            else:
-                code += f"    {unconstrained_reference} = unconstrained_parameter_vector[..., {offset}]\n"
-
-            variable = self.parameter_variables[name]
-            lower = variable.lower
-            upper = variable.upper
-
-            if lower > float("-inf") or upper < float("inf"):
-                if lower > float("-inf") and upper == float("inf"):
-                    code += f"    {constrained_reference}, constraints_jacobian_adjustment = rat.constraints.lower({unconstrained_reference}, {lower})\n"
-                elif lower == float("inf") and upper < float("inf"):
-                    code += f"    {constrained_reference}, constraints_jacobian_adjustment = rat.constraints.upper({unconstrained_reference}, {upper})\n"
-                elif lower > float("-inf") and upper < float("inf"):
-                    code += f"    {constrained_reference}, constraints_jacobian_adjustment = rat.constraints.finite({unconstrained_reference}, {lower}, {upper})\n"
-
-                code += "    total += jax.numpy.sum(constraints_jacobian_adjustment)\n"
-            else:
-                code += f"    {constrained_reference} = {unconstrained_reference}\n"
-
-            if size is not None and size != variable.padded_size():
-                code += f"    if pad:\n"
-                code += f"        {constrained_reference} = jax.numpy.pad({constrained_reference}, (0, {variable.padded_size() - size}))\n"
-        code += "\n    return total, parameters\n"
-
-        ordered_expr_trees = list(reversed(self.expr_tree_list))
-
-        # Transform parameters
-        code += "\ndef transform_parameters(data, subscripts, first_in_group_indicators, parameters):"
-        for top_expr in ordered_expr_trees:
-            match top_expr:
-                case ops.Assignment(lhs, rhs):
-                    # Assume left hand side is a parameter
-                    lhs_key = top_expr.lhs.get_key()
-                    code += f"\n    # {lhs_key}\n"
-
-                    lhs_size = lhs.variable.size()
-                    lhs_is_scalar = lhs_size == None
-                    lhs_used_in_rhs = False
-                    need_first_in_group_indicator = False
-                    if not lhs_is_scalar:
-                        carry_size = 0
-                        # Because left hand side is parameter, we only need to search for parameters on the right hand side
-                        for symbol in ops.search_tree(rhs, ops.Param):
-                            symbol_key = symbol.get_key()
-
-                            # Save an indicator if the left hand side is used on the right hand side
-                            if symbol_key == lhs_key:
-                                lhs_used_in_rhs = True
-
-                                # Figure out the carry size needed
-                                shifts = symbol.subscript.shifts
-
-                                carry_size = max(carry_size, shifts[-1])
-
-                                # There will be groupings if there is more than one shift
-                                if len(shifts) > 1:
-                                    need_first_in_group_indicator = True
-
-                    if lhs_used_in_rhs:
-                        # We need to scan in this case
-                        scan_function_name = f"scan_function_{self._get_unique_identifier()}"
-                        zero_tuple_string = f"({','.join(['0.0'] * carry_size)})"
-                        carry_strings = [f"carry{n}" for n in range(carry_size)]
-                        carry_tuple_string = f"({','.join(carry_strings)})"
-                        next_carry_tuple_string = f"({','.join(['next_value'] + carry_strings[:-1])})"
-
-                        code += f"    def {scan_function_name}(carry, index):\n"
-                        if need_first_in_group_indicator:
-                            code += f"        carry = jax.lax.cond(first_in_group_indicators['{lhs_key}'][index], lambda : {zero_tuple_string}, lambda : carry)\n"
-                        code += f"        {carry_tuple_string} = carry\n"
-                        code += f"        next_value = {top_expr.rhs.code(scalar = True)}\n"
-                        code += f"        return {next_carry_tuple_string}, next_value\n"
-
-                        code += f"\n    _, parameters['{lhs_key}'] = jax.lax.scan({scan_function_name}, {zero_tuple_string}, jax.numpy.arange({lhs_size}))\n"
+                # if param_key == primary_variable_key:
+                #     continue
+
+                # fold shift values into IntegerConstant if it's an expression instead of a constant
+                folded_shift_exprs: List[ast.IntegerConstant] = []
+                for shift_expr in param.subscript.shifts:
+                    if isinstance(shift_expr, ast.IntegerConstant):
+                        folded_shift_exprs.append(shift_expr)
                     else:
-                        code += f"    parameters['{lhs_key}'] = {top_expr.rhs.code()}\n"
-        code += "\n    return parameters\n"
+                        try:
+                            shift_code_visitor = codegen_backends.BaseVisitor(self.symbol_table)
+                            shift_expr.accept(shift_code_visitor)
+                            folded_integerconstant = ast.IntegerConstant(value=int(eval(shift_code_visitor.expression_string)))
+                            folded_shift_exprs.append(folded_integerconstant)
 
-        # Generate code for evaluating densities
-        code += f"\ndef evaluate_densities(data, subscripts, parameters):\n" f"    target = 0.0\n"
+                        except Exception as e:
+                            error_msg = f"Failed evaluating shift amounts for parameter {param_key}. Shift amount expressions must be an expression which can be evaluated at compile-time."
+                            raise CompileError(error_msg, self.model_code_string, shift_expr.line_index, shift_expr.column_index) from e
 
-        for top_expr in ordered_expr_trees:
-            if isinstance(top_expr, ops.Distr):
-                code += f"    target += jax.numpy.sum({top_expr.code()})\n"
-        code += "\n    return target"
+                # If there is shift, set pad needed to True
+                for integer_constant in folded_shift_exprs:
+                    if integer_constant.value != 0:
+                        self.symbol_table.lookup(param_key).pad_needed = True
+                        break
+                param.subscript.shifts = folded_shift_exprs
+                ###########
 
-        return code
+                # Determine type of the variable. If its name exists in the input dataframe, it's a data variable.
+                # Param is the default value. Assigned params are upserted when the compiler identifies it being used on
+                # the LHS of an assignment.
+                try:
+                    var_type = self.symbol_table.lookup(param_key).variable_type
+                except KeyError:
+                    var_type = VariableType.DATA if param_key in self.data_df_columns else VariableType.PARAM
+                ###########
 
-    def compile(
-        self,
-    ) -> Tuple[
-        Dict[str, variables.Data],
-        Dict[str, variables.Param],
-        Dict[str, variables.AssignedParam],
-        Dict[str, variables.SubscriptUse],
-        Dict[str, numpy.ndarray],
-        str,
-    ]:
-        if self.parameter_variables:
-            raise Exception("Compiler.compile() may be invoked only once per instance. Create a new instance to recompile.")
+                # We "resolve" subscripts, so every single parameters' subscripts can be described with only the input
+                # dataframe's columns. The symbol table holds the "true" subscript columns for each parameter.
+                n_subscripts = len(param.subscript.names)
+                subscript_list = [[] for _ in range(n_subscripts)]
 
+                variable_subscript_alias: List[str] = []  # This list holds the actual column names of the base DF
+                for index in range(n_subscripts):
+                    subscript_column = param.subscript.names[index]
+                    subscript_name = subscript_column.name
+
+                    if subscript_name in subscript_aliases:
+                        subscript_list[index].extend(subscript_aliases[subscript_name])
+                    elif subscript_name in allowed_subscript_names:
+                        subscript_list[index].append(subscript_name)
+                    else:
+                        raise CompileError(
+                            f"Subscript name '{subscript_name}' not in scope.",
+                            self.model_code_string,
+                            subscript_column.line_index,
+                            subscript_column.column_index,
+                        )
+                    variable_subscript_alias.append(subscript_name)
+
+                # store the "true" subscripts
+                # ex) sub_1 = {sub_2, sub_3, sub_4}
+                # then the true subscripts of param[sub_1, x] is [(sub_2, x), (sub_3, x), (sub_4, x)]
+                unrolled_subscripts = [tuple(x) for x in itertools.product(*subscript_list)]
+
+                self.symbol_table.upsert(
+                    param_key, var_type, subscripts=set(unrolled_subscripts), subscript_alias=tuple(variable_subscript_alias)
+                )
+
+    def codegen(self):
+        self.generated_code = ""
+        self.generated_code += "# A rat model\n\n"
+        self.generated_code += "import rat.constraints\n"
+        self.generated_code += "import rat.math\n"
+        self.generated_code += "import jax\n\n"
+
+        self.generated_code += f"unconstrained_parameter_size = {self.symbol_table.unconstrained_param_count}\n\n"
+
+        self.codegen_constrain_parameters()
+
+        self.codegen_transform_parameters()
+
+        self.codegen_evaluate_densities()
+
+    def codegen_constrain_parameters(self):
+        self.generated_code += "def constrain_parameters(unconstrained_parameter_vector, pad=True):\n"
+        self.generated_code += "    unconstrained_parameters = {}\n"
+        self.generated_code += "    parameters = {}\n"
+        self.generated_code += "    jacobian_adjustments = 0.0\n"
+
+        for variable_name, record in self.symbol_table.symbol_dict.items():
+            if record.variable_type != VariableType.PARAM:
+                continue
+
+            unconstrained_reference = f"unconstrained_parameters['{variable_name}']"
+            constrained_reference = f"parameters['{variable_name}']"
+
+            self.generated_code += "\n"
+            self.generated_code += f"    # Param: {variable_name}, lower: {record.constraint_lower}, upper: {record.constraint_upper}\n"
+
+            # This assumes that unconstrained parameter indices for a parameter is allocated in a contiguous fashion.
+            index_string = (
+                f"{record.unconstrained_vector_start_index} : {record.unconstrained_vector_end_index + 1}"
+                if record.unconstrained_vector_start_index != record.unconstrained_vector_end_index
+                else f"{record.unconstrained_vector_start_index}"
+            )
+            self.generated_code += f"    {unconstrained_reference} = unconstrained_parameter_vector[..., {index_string}]\n"
+
+            if record.constraint_lower > float("-inf") or record.constraint_upper < float("inf"):
+                if record.constraint_lower > float("-inf") and record.constraint_upper == float("inf"):
+                    self.generated_code += f"    {constrained_reference}, constraints_jacobian_adjustment = rat.constraints.lower({unconstrained_reference}, {record.constraint_lower})\n"
+                elif record.constraint_lower == float("inf") and record.constraint_upper < float("inf"):
+                    self.generated_code += f"    {constrained_reference}, constraints_jacobian_adjustment = rat.constraints.upper({unconstrained_reference}, {record.constraint_upper})\n"
+                elif record.constraint_lower > float("-inf") and record.constraint_upper < float("inf"):
+                    self.generated_code += f"    {constrained_reference}, constraints_jacobian_adjustment = rat.constraints.finite({unconstrained_reference}, {record.constraint_lower}, {record.constraint_upper})\n"
+
+                self.generated_code += "    jacobian_adjustments += jax.numpy.sum(constraints_jacobian_adjustment)\n"
+            else:
+                self.generated_code += f"    {constrained_reference} = {unconstrained_reference}\n"
+
+            if record.pad_needed:
+                self.generated_code += "    if pad:\n"
+                self.generated_code += f"        {constrained_reference} = jax.numpy.pad({constrained_reference}, (0, 1))\n"
+
+        self.generated_code += "\n"
+        self.generated_code += "    return jacobian_adjustments, parameters\n\n"
+
+    def codegen_transform_parameters(self):
+        self.generated_code += "def transform_parameters(data, subscripts, first_in_group_indicators, parameters)\n"
+
+        for top_expr in self.expr_tree_list:
+            if not isinstance(top_expr, ast.Assignment):
+                continue
+
+            codegen_visitor = codegen_backends.TransformedParametersVisitor(self.symbol_table, self._get_primary_symbol_from_statement(top_expr))
+            top_expr.accept(codegen_visitor)
+            self.generated_code += "    "
+            self.generated_code += codegen_visitor.expression_string
+
+        self.generated_code += "\n"
+        self.generated_code += "    return parameters\n\n"
+
+    def codegen_evaluate_densities(self):
+        self.generated_code += "def evaluate_densities(data, subscripts, parameters):\n"
+        self.generated_code += "    target = 0.0\n"
+
+        for top_expr in self.expr_tree_list:
+            if not isinstance(top_expr, ast.Distr):
+                continue
+
+            codegen_visitor = codegen_backends.EvaluateDensityVisitor(self.symbol_table, self._get_primary_symbol_from_statement(top_expr))
+            top_expr.accept(codegen_visitor)
+            self.generated_code += f"    target += jax.numpy.sum({codegen_visitor.expression_string})\n"
+
+        self.generated_code += "\n"
+        self.generated_code += "    return target\n"
+
+    def compile(self):
         self._identify_primary_symbols()
 
-        # build the parameter_base_df for each variable
-        self._build_base_df()
+        self.build_symbol_table()
 
-        # build variable.Subscript for each parameter
-        logging.debug("now printing parameter_base_df")
-        for var_name, subscript_names in self.parameter_subscript_names.items():
-            logging.debug(var_name)
-            logging.debug(self.parameter_base_df[var_name])
-            self.variable_subscripts[var_name] = variables.Subscript(self.parameter_base_df[var_name], subscript_names)
-        logging.debug("done printing parameter_base_df")
+        self.symbol_table.build_base_dataframes(self.data_df)
 
-        # build variable.Subscript for each data
-        logging.debug("now printing data_base_df")
-        for var_name, base_df in self.data_base_df.items():
-            logging.debug(var_name)
-            self.variable_subscripts[var_name] = variables.Subscript(base_df, [])
-        logging.debug("done printing data_base_df")
+        print(self.symbol_table)
 
-        logging.debug("now printing variable_subscripts")
-        for key, val in self.variable_subscripts.items():
-            logging.debug(key)
-            logging.debug(val.base_df)
-            logging.info(f"Subscript data for parameter {key}:")
-            val.log_summary(logging.INFO)
-            logging.info("-----")
-        logging.debug("done printing variable_subscripts")
-
-        # run seconds pass, which binds subscripts and does codegen
-        self._build_variables()
-
-        # search for and prepare for assignments that must be handled
-        # with scans (these are ones where the left hand side symbol
-        # appears on the right hand side)
-        self._prepare_for_scans()
-
-        return (
-            self.data_variables,
-            self.parameter_variables,
-            self.assigned_parameter_variables,
-            self.subscript_use_variables,
-            self.first_in_group_indicators,
-            self._generate_python_source(),
-        )
+        self.codegen()
